+++ Covenant/API/Models/PowerShellLauncher.cs
@@ -24,7 +24,7 @@ namespace Covenant.API.Models
         /// </summary>
         /// <param name="type">Possible values include: 'Wmic', 'Regsvr32',
         /// 'Mshta', 'Cscript', 'Wscript', 'PowerShell', 'Binary', 'MSBuild',
-        /// 'InstallUtil'</param>
+        /// 'InstallUtil', 'ShellCode'</param>
         /// <param name="dotNetVersion">Possible values include: 'Net35',
         /// 'Net40', 'NetCore31'</param>
         /// <param name="runtimeIdentifier">Possible values include: 'win_x64',
@@ -114,7 +114,7 @@ namespace Covenant.API.Models
         /// <summary>
         /// Gets or sets possible values include: 'Wmic', 'Regsvr32', 'Mshta',
         /// 'Cscript', 'Wscript', 'PowerShell', 'Binary', 'MSBuild',
-        /// 'InstallUtil'
+        /// 'InstallUtil', 'ShellCode'
         /// </summary>
         [JsonProperty(PropertyName = "type")]
         public LauncherType? Type { get; set; }