+++ Covenant/Core/CovenantService.cs
@@ -306,6 +306,10 @@ namespace Covenant.Core
         Task<BinaryLauncher> GenerateBinaryLauncher();
         Task<BinaryLauncher> GenerateBinaryHostedLauncher(HostedFile file);
         Task<BinaryLauncher> EditBinaryLauncher(BinaryLauncher launcher);
+        Task<ShellCodeLauncher> GetShellCodeLauncher();
+        Task<ShellCodeLauncher> GenerateShellCodeLauncher();
+        Task<ShellCodeLauncher> GenerateShellCodeHostedLauncher(HostedFile file);
+        Task<ShellCodeLauncher> EditShellCodeLauncher(ShellCodeLauncher launcher);
         Task<PowerShellLauncher> GetPowerShellLauncher();
         Task<PowerShellLauncher> GeneratePowerShellLauncher();
         Task<PowerShellLauncher> GeneratePowerShellHostedLauncher(HostedFile file);
@@ -4280,6 +4284,95 @@ public static class Task
             await _context.SaveChangesAsync();
             // _notifier.OnEditLauncher(this, matchingLauncher);
             return await this.GetBinaryLauncher();
+        }
+
+        public async Task<ShellCodeLauncher> GetShellCodeLauncher()
+        {
+            ShellCodeLauncher launcher = (ShellCodeLauncher)await _context.Launchers.FirstOrDefaultAsync(S => S.Type == LauncherType.ShellCode);
+            if (launcher == null)
+            {
+                throw new ControllerNotFoundException($"NotFound - ShellCodeLauncher");
+            }
+            return launcher;
+        }
+
+        public async Task<ShellCodeLauncher> GenerateShellCodeLauncher()
+        {
+            ShellCodeLauncher launcher = await this.GetShellCodeLauncher();
+            Listener listener = await this.GetListener(launcher.ListenerId);
+            ImplantTemplate template = await this.GetImplantTemplate(launcher.ImplantTemplateId);
+            Profile profile = await this.GetProfile(listener.ProfileId);
+
+            if (!template.CompatibleListenerTypes.Select(LT => LT.Id).Contains(listener.ListenerTypeId))
+            {
+                throw new ControllerBadRequestException($"BadRequest - ListenerType not compatible with chosen ImplantTemplate");
+            }
+
+            Grunt grunt = new Grunt
+            {
+                ListenerId = listener.Id,
+                Listener = listener,
+                ImplantTemplateId = template.Id,
+                ImplantTemplate = template,
+                SMBPipeName = launcher.SMBPipeName,
+                ValidateCert = launcher.ValidateCert,
+                UseCertPinning = launcher.UseCertPinning,
+                Delay = launcher.Delay,
+                JitterPercent = launcher.JitterPercent,
+                ConnectAttempts = launcher.ConnectAttempts,
+                KillDate = launcher.KillDate,
+                DotNetVersion = launcher.DotNetVersion,
+                RuntimeIdentifier = launcher.RuntimeIdentifier
+            };
+
+            await _context.Grunts.AddAsync(grunt);
+            await _context.SaveChangesAsync();
+            await _notifier.NotifyCreateGrunt(this, grunt);
+
+            launcher.GetLauncher(
+                this.GruntTemplateReplace(template.StagerCode, template, grunt, listener, profile),
+                CompileGruntCode(template.StagerCode, template, grunt, listener, profile, launcher),
+                grunt,
+                template
+            );
+            _context.Launchers.Update(launcher);
+            await _context.SaveChangesAsync();
+            // _notifier.OnEditLauncher(this, launcher);
+            return await this.GetShellCodeLauncher();
+        }
+
+        public async Task<ShellCodeLauncher> GenerateShellCodeHostedLauncher(HostedFile file)
+        {
+            ShellCodeLauncher launcher = await this.GetShellCodeLauncher();
+            Listener listener = await this.GetListener(launcher.ListenerId);
+            HostedFile savedFile = await this.GetHostedFile(file.Id);
+            string hostedLauncher = launcher.GetHostedLauncher(listener, savedFile);
+            _context.Launchers.Update(launcher);
+            await _context.SaveChangesAsync();
+            // _notifier.OnEditLauncher(this, launcher);
+            return await this.GetShellCodeLauncher();
+        }
+
+        public async Task<ShellCodeLauncher> EditShellCodeLauncher(ShellCodeLauncher launcher)
+        {
+            ShellCodeLauncher matchingLauncher = await this.GetShellCodeLauncher();
+            Listener listener = await this.GetListener(launcher.ListenerId);
+            matchingLauncher.ListenerId = listener.Id;
+            matchingLauncher.ImplantTemplateId = launcher.ImplantTemplateId;
+            matchingLauncher.DotNetVersion = launcher.DotNetVersion;
+            matchingLauncher.RuntimeIdentifier = launcher.RuntimeIdentifier;
+            matchingLauncher.SMBPipeName = launcher.SMBPipeName;
+            matchingLauncher.ValidateCert = launcher.ValidateCert;
+            matchingLauncher.UseCertPinning = launcher.UseCertPinning;
+            matchingLauncher.Delay = launcher.Delay;
+            matchingLauncher.JitterPercent = launcher.JitterPercent;
+            matchingLauncher.ConnectAttempts = launcher.ConnectAttempts;
+            matchingLauncher.KillDate = launcher.KillDate;
+            matchingLauncher.LauncherString = launcher.LauncherString;
+            _context.Launchers.Update(matchingLauncher);
+            await _context.SaveChangesAsync();
+            // _notifier.OnEditLauncher(this, matchingLauncher);
+            return await this.GetShellCodeLauncher();
         }
 
         public async Task<PowerShellLauncher> GetPowerShellLauncher()